Move Timer in new package jflex.performance
