+++ jflex/src/main/java/jflex/LexGenerator.java
@@ -17,6 +17,7 @@ import java.io.InputStreamReader;
 import java.io.Reader;
 import java.nio.file.Files;
 import java.nio.file.Paths;
+import jflex.performance.Timer;
 
 /**
  * This is the generator of JFlex, controlling the scanner generation process.
