+++ jflex/src/main/java/jflex/Out.java
@@ -15,6 +15,7 @@ import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import java.io.OutputStream;
+import jflex.performance.Timer;
 import jflex.unicode.UnicodeProperties;
 
 /**
