+++ jflex/src/main/jflex/LexScan.flex
@@ -18,6 +18,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.HashMap;
+import jflex.performance.Timer;
 import jflex.unicode.UnicodeProperties;
 
 %%
@@ -97,10 +98,10 @@ import jflex.unicode.UnicodeProperties;
   String functionName;
   String tokenType;
   String visibility = "public";
-    
+
   List<String> ctorArgs = new ArrayList<String>();
   List<String> ctorTypes = new ArrayList<String>();
-    
+
   LexicalStates states = new LexicalStates();
 
   List<Action> actions = new ArrayList<Action>();
@@ -114,7 +115,7 @@ import jflex.unicode.UnicodeProperties;
   // CharClasses.init() is delayed until UnicodeProperties.init() has been called,
   // since the max char code won't be known until then.
   private CharClasses charClasses = new CharClasses();
-  
+
   public CharClasses getCharClasses() {
     return charClasses;
   }
@@ -180,11 +181,11 @@ import jflex.unicode.UnicodeProperties;
 
     return a.toString()+", "+b.toString();
   }
-  
+
   public UnicodeProperties getUnicodeProperties() {
     return unicodeProperties;
   }
-  
+
   private void populateDefaultVersionUnicodeProperties() {
     try {
       unicodeProperties = new UnicodeProperties();
@@ -195,7 +196,7 @@ import jflex.unicode.UnicodeProperties;
     charClasses.init
       (Options.jlex ? 127 : unicodeProperties.getMaximumCodePoint(), this);
   }
-  
+
   private void includeFile(String filePath) {
     File f = new File(file.getParentFile(), filePath);
     if ( !f.canRead() )
